feat: Mudanças de validação e lógica
- Adiciona a dependência fastify-type-provider-zod para fornecer integração com Zod.
- Adiciona o provedor de tipo Zod para Fastify.
- Adiciona autenticação e autorização para atualizar e excluir comentários, garantindo que apenas usuários autorizados possam realizar essas ações.
- Atualiza a lógica de atualização e remoção de posts para incorporar informações do usuário na autorização e auditoria.
- Adiciona classe de erro personalizada para padronizar o tratamento de erros na aplicação.
- Adiciona classes de erro personalizadas para representar erros HTTP comuns, como NotFound, Unauthorized e Forbidden.
- Implementa tratamento de erros centralizado para respostas de erros da aplicação.
- Renomeia o arquivo de configuração JWT para o arquivo de middleware de verificação JWT.

- Implementa tratamento de erro específico e lança exceções de `AppError` para ações não autorizadas ou permissões negadas.
- Adiciona esquemas Zod para validação e tipagem de DTOs de comentários.
- Use Zod para definir e tipar esquemas de DTOs de postagem.

- Define esquemas Zod para validação e tipagem de dados de usuário nas operações de criação, atualização, resposta e exclusão.
- Atualiza o middleware de verificação JWT para um diretório dedicado.
- Adiciona documentação Swagger para endpoints de busca e atualização de posts e move middleware de verificação JWT.
- Implemente a configuração do Fastify com Zod, registre manipuladores de erro e configure rotas de documentação Swagger/OpenAPI.

- Implementa autorizações e tratamento de erros nas operações de atualização e exclusão de comentários.
- Implementa a lógica de autorização e permissões para atualizar e excluir posts, além de adicionar tratamento de erros e restrição de campos retornados na busca de posts.

+++ fastify/src/service/comment.ts
@@ -1,3 +1,5 @@
+import { AppError } from "../core/errors/AppError";
+import { checkPermission } from "../core/permissions/permissions";
 import prisma from "../core/utils/prismaClient";
 
 import { CreateCommentDTO, UpdateCommentDTO } from "../dto/comment";
@@ -16,13 +18,31 @@ export const commentService = () => ({
             data
         });
     },
-    update: async (id: string, data: UpdateCommentDTO) => {
+    update: async (id: string,user:{ role: string, id: string }, data: UpdateCommentDTO) => {
+        const comment = await prisma.comment.findUnique({
+            where: { id }
+        });
+
+        if (!comment) {
+            throw new AppError('Comment not found', 404);
+        }
+
+        checkPermission(user.role,"update","comment", user.id, comment!.authorId);
         return await prisma.comment.update({
             where: { id },
             data
         });
     },
-    delete: async (id: string) => {
+    delete: async (id: string, user:{ role: string, id: string }) => {
+         const comment = await prisma.comment.findUnique({
+            where: { id }
+        });
+        console.log("Comment achado", comment);
+        if (!comment) {
+            throw new AppError('Comment not found', 404);
+        }
+
+        checkPermission(user.role,"delete","comment", user.id, comment!.authorId);
         return await prisma.comment.delete({
             where: { id}
         });
